refactor(constant): 패키지 변경
