+++ src/main/java/store/constant/ProductType.java
@@ -1,4 +1,4 @@
-package store.model.domain;
+package store.constant;
 
 public enum ProductType {
     REGULAR,
